+++ apps/api/src/app/symbol/interfaces/lookup-item.interface.ts
@@ -1,6 +1,8 @@
-import { DataSource } from '@prisma/client';
+import { AssetClass, AssetSubClass, DataSource } from '@prisma/client';
 
 export interface LookupItem {
+  assetClass: AssetClass;
+  assetSubClass: AssetSubClass;
   currency: string;
   dataSource: DataSource;
   name: string;