+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added the asset and asset sub class to the search functionality
 - Added the subscription expiration date to the users table of the admin control panel
 
 ### Changed